튜토리얼 퀘스트 처리

+++ GP_Server/Source/Game/DataTable/QuestTable.h
@@ -7,7 +7,7 @@ struct QuestData
 	int32 ExpReward;
 	int32 GoldReward;
 	QuestType NextQuestID = QuestType::NONE;
-	uint8 RewarditemType;
+	int32 RewarditemType;
 };
 
 class QuestTable
